[chore] Separated near-primitives dependencies into features (#11597)
This pull request decreases near-primitive dependencies count even more
by separating part of functionality into features.

I have created three new features:
* `rand` - extracts rand, rand-chacha crates from the compilation. This
is common practice to split by this type.
* `solomon` - extracts reed-solomon-erasure, libc, lru, and some
outdated subdeps (e.g. hashbrowns)
* `clock` - removes `tokio` from dependency list that is nice :)

Moreover, I have featured gated rand in near-crypto following the same
logic :)

Final results: near-primitives have 107 dependencies (-23) using a
feature-less version
