+++ .gitignore
@@ -68,3 +68,4 @@ rusty-tags.vi
 costs-*.txt
 names-to-stats.txt
 data_dump_*.bin
+.venv
