+++ core/async/Cargo.toml
@@ -19,13 +19,13 @@ once_cell.workspace = true
 serde.workspace = true
 serde_json.workspace = true
 time.workspace = true
-tokio = {workspace = true, features = ["rt", "macros"]}
+tokio = { workspace = true, features = ["rt", "macros"] }
 tracing.workspace = true
 
 near-async-derive.workspace = true
 near-o11y.workspace = true
 near-performance-metrics.workspace = true
-near-time.workspace = true
+near-time = { workspace = true, features = ["clock"] }
 
 [dev-dependencies]
 derive-enum-from-into.workspace = true
