+++ core/primitives/Cargo.toml
@@ -22,13 +22,13 @@ chrono.workspace = true
 derive_more.workspace = true
 easy-ext.workspace = true
 hex.workspace = true
-itertools.workspace = true
+itertools = { workspace = true, optional = true }
 num-rational.workspace = true
 once_cell.workspace = true
 primitive-types.workspace = true
-rand.workspace = true
-rand_chacha.workspace = true
-reed-solomon-erasure.workspace = true
+rand = { workspace = true, optional = true }
+rand_chacha = { workspace = true, optional = true }
+reed-solomon-erasure = { workspace = true, optional = true }
 serde.workspace = true
 serde_json.workspace = true
 serde_with.workspace = true
@@ -51,6 +51,9 @@ near-parameters.workspace = true
 [features]
 sandbox = []
 test_features = []
+solomon = ["reed-solomon-erasure", "itertools"]
+rand = ["dep:rand", "rand_chacha", "near-crypto/rand", "itertools"]
+clock = ["near-time/clock"]
 dump_errors_schema = ["near-rpc-error-macro/dump_errors_schema"]
 protocol_feature_fix_staking_threshold = [
   "near-primitives-core/protocol_feature_fix_staking_threshold",
@@ -69,6 +72,7 @@ nightly = [
   "near-fmt/nightly",
   "near-parameters/nightly",
   "near-primitives-core/nightly",
+  "near-primitives/nightly",
   "nightly_protocol",
   "protocol_feature_fix_contract_loading_cost",
   "protocol_feature_fix_staking_threshold",
@@ -80,6 +84,7 @@ nightly_protocol = [
   "near-fmt/nightly_protocol",
   "near-parameters/nightly_protocol",
   "near-primitives-core/nightly_protocol",
+  "near-primitives/nightly_protocol",
 ]
 
 statelessnet_protocol = ["near-primitives-core/statelessnet_protocol"]
@@ -90,12 +95,15 @@ new_epoch_sync = []
 calimero_zero_storage = []
 
 [dev-dependencies]
+chrono = { workspace = true, features = ["clock"] }
+near-primitives = { workspace = true, features = ["clock", "solomon", "rand"] }
 assert_matches.workspace = true
 bencher.workspace = true
 bolero.workspace = true
 insta.workspace = true
 expect-test.workspace = true
 regex.workspace = true
+
 
 [[bench]]
 name = "serialization"
