+++ core/primitives/src/block.rs
@@ -6,23 +6,18 @@ use crate::block_body::{BlockBody, BlockBodyV1, ChunkEndorsementSignatures};
 pub use crate::block_header::*;
 use crate::challenge::{Challenges, ChallengesResult};
 use crate::checked_feature;
-use crate::congestion_info::{BlockCongestionInfo, CongestionInfo, ExtendedCongestionInfo};
+use crate::congestion_info::{BlockCongestionInfo, ExtendedCongestionInfo};
 use crate::hash::{hash, CryptoHash};
 use crate::merkle::{merklize, verify_path, MerklePath};
 use crate::num_rational::Rational32;
-use crate::sharding::{
-    ChunkHashHeight, EncodedShardChunk, ShardChunk, ShardChunkHeader, ShardChunkHeaderV1,
-};
-use crate::types::{Balance, BlockHeight, EpochId, Gas, NumBlocks, StateRoot};
-use crate::validator_signer::{EmptyValidatorSigner, ValidatorSigner};
+use crate::sharding::{ChunkHashHeight, ShardChunkHeader, ShardChunkHeaderV1};
+use crate::types::{Balance, BlockHeight, EpochId, Gas, NumBlocks};
+use crate::validator_signer::ValidatorSigner;
 use crate::version::{ProtocolVersion, SHARD_CHUNK_HEADER_UPGRADE_VERSION};
 use borsh::{BorshDeserialize, BorshSerialize};
 use near_crypto::Signature;
-use near_primitives_core::types::ShardId;
-use near_primitives_core::version::ProtocolFeature;
 use near_time::Utc;
 use primitive_types::U256;
-use reed_solomon_erasure::galois_8::ReedSolomon;
 use std::collections::BTreeMap;
 use std::ops::Index;
 use std::sync::Arc;
@@ -91,14 +86,15 @@ pub enum Block {
     BlockV4(Arc<BlockV4>),
 }
 
+#[cfg(feature = "solomon")]
 pub fn genesis_chunks(
-    state_roots: Vec<StateRoot>,
-    shard_ids: &[ShardId],
+    state_roots: Vec<crate::types::StateRoot>,
+    shard_ids: &[crate::types::ShardId],
     initial_gas_limit: Gas,
     genesis_height: BlockHeight,
     genesis_protocol_version: ProtocolVersion,
-) -> Vec<ShardChunk> {
-    let rs = ReedSolomon::new(1, 2).unwrap();
+) -> Vec<crate::sharding::ShardChunk> {
+    let rs = reed_solomon_erasure::galois_8::ReedSolomon::new(1, 2).unwrap();
     let state_roots = if state_roots.len() == shard_ids.len() {
         state_roots
     } else {
@@ -106,15 +102,15 @@ pub fn genesis_chunks(
         std::iter::repeat(state_roots[0]).take(shard_ids.len()).collect()
     };
 
-    let congestion_info = ProtocolFeature::CongestionControl
+    let congestion_info = near_primitives_core::version::ProtocolFeature::CongestionControl
         .enabled(genesis_protocol_version)
-        .then_some(CongestionInfo::default());
+        .then_some(crate::congestion_info::CongestionInfo::default());
 
     shard_ids
         .into_iter()
         .zip(state_roots)
         .map(|(&shard_id, state_root)| {
-            let (encoded_chunk, _) = EncodedShardChunk::new(
+            let (encoded_chunk, _) = crate::sharding::EncodedShardChunk::new(
                 CryptoHash::default(),
                 state_root,
                 CryptoHash::default(),
@@ -130,7 +126,7 @@ pub fn genesis_chunks(
                 &[],
                 CryptoHash::default(),
                 congestion_info,
-                &EmptyValidatorSigner::default().into(),
+                &crate::validator_signer::EmptyValidatorSigner::default().into(),
                 genesis_protocol_version,
             )
             .expect("Failed to decode genesis chunk");
