+++ core/primitives/src/sharding.rs
@@ -2,7 +2,6 @@ use crate::congestion_info::CongestionInfo;
 use crate::hash::{hash, CryptoHash};
 use crate::merkle::{combine_hash, merklize, verify_path, MerklePath};
 use crate::receipt::Receipt;
-use crate::reed_solomon::reed_solomon_encode;
 use crate::transaction::SignedTransaction;
 use crate::types::validator_stake::{ValidatorStake, ValidatorStakeIter, ValidatorStakeV1};
 use crate::types::{Balance, BlockHeight, Gas, MerkleHash, ShardId, StateRoot};
@@ -11,7 +10,6 @@ use crate::version::{ProtocolFeature, ProtocolVersion, SHARD_CHUNK_HEADER_UPGRAD
 use borsh::{BorshDeserialize, BorshSerialize};
 use near_crypto::Signature;
 use near_fmt::AbbrBytes;
-use reed_solomon_erasure::galois_8::ReedSolomon;
 use std::cmp::Ordering;
 use std::sync::Arc;
 use tracing::debug_span;
@@ -1021,13 +1019,14 @@ impl EncodedShardChunk {
         TransactionReceipt::try_from_slice(&encoded_data)
     }
 
+    #[cfg(feature = "solomon")]
     pub fn new(
         prev_block_hash: CryptoHash,
         prev_state_root: StateRoot,
         prev_outcome_root: CryptoHash,
         height: BlockHeight,
         shard_id: ShardId,
-        rs: &ReedSolomon,
+        rs: &reed_solomon_erasure::galois_8::ReedSolomon,
         prev_gas_used: Gas,
         gas_limit: Gas,
         prev_balance_burnt: Balance,
@@ -1040,7 +1039,7 @@ impl EncodedShardChunk {
         signer: &ValidatorSigner,
         protocol_version: ProtocolVersion,
     ) -> Result<(Self, Vec<MerklePath>), std::io::Error> {
-        let (transaction_receipts_parts, encoded_length) = reed_solomon_encode(
+        let (transaction_receipts_parts, encoded_length) = crate::reed_solomon::reed_solomon_encode(
             rs,
             TransactionReceipt(transactions, prev_outgoing_receipts.to_vec()),
         );
